Add files via upload
A robotic vehicle with internally fed intelligence that can change its course when an obstacle is hindering its way. In order to prevent physical harm to the robot. The robotic vehicle moves by making use of ultrasonic sensors to ignore hinderances on its path. To carry out the needed procedure, an AT mega 328P
microcontroller is employed. The micro-controller receives an instruction from the ultrasonic sensor, which uses sound waves to identify any obstacles in front of it. The microcontroller activates the motors to command the machine to move in another direction based on the input signal it receives which are instructed and given input to it via motor driver.